requirements.txt
altair==6.0.0
feedparser==6.0.11
fpdf==1.7.2
google-genai==1.57.0
schedule==1.2.2
python-dotenv==1.0.1

# Core dependencies (let pip resolve compatible versions)
annotated-doc
annotated-types
anyio>=4.8.0,<5.0.0
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
distro
fastapi
gitdb
GitPython
h11
httpcore
httpx
idna
Jinja2
jiter
jsonschema
jsonschema-specifications
lxml
MarkupSafe
narwhals
numpy
openai
packaging
pandas
pillow
protobuf
pyarrow
pydantic
pydantic_core
pydeck
python-dateutil
pytz
referencing
requests
rpds-py
six
smmap
sniffio
soupsieve
starlette
streamlit==1.52.2
tenacity
groq>=0.18.0
openai>=1.0.0
toml
tornado
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
